backend/requirements.txt
appdirs==1.4.4
#boto3==1.26.74
#botocore==1.29.74
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.5
cryptography==38.0.4
cycler==0.11.0
fonttools==4.25.0
greenlet==2.0.2
idna==3.4
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
numexpr==2.8.4
numpy
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pip==22.3.1
pooch==1.6.0
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
setuptools==65.6.3
six==1.16.0
#SQLAlchemy==1.4.39
threadpoolctl==3.1.0
tornado==6.2
urllib3==1.26.14
wheel==0.38.4
yagmail==0.15.293

fastapi~=0.92.0
uvicorn~=0.20.0
SQLAlchemy~=2.0.4
python-dotenv~=0.21.1
pydantic~=1.10.5
boto3~=1.26.75
botocore~=1.29.75
pika~=1.3.1